Revert "Submodule lowering"
