+++ src/CodeGen_LLVM.h
@@ -128,7 +128,6 @@ protected:
     static bool llvm_Mips_enabled;
     static bool llvm_PowerPC_enabled;
 
-    const Module *input_module;
     std::unique_ptr<llvm::Module> module;
     llvm::Function *function;
     llvm::LLVMContext *context;
@@ -234,8 +233,8 @@ protected:
     /** Take an llvm Value representing a pointer to a buffer_t,
      * and populate the symbol table with its constituent parts.
      */
-    void push_buffer(const std::string &name, int dimensions, llvm::Value *buffer, bool global = false);
-    void pop_buffer(const std::string &name, int dimensions, bool global = false);
+    virtual void push_buffer(const std::string &name, int dimensions, llvm::Value *buffer);
+    virtual void pop_buffer(const std::string &name, int dimensions);
 
     /** Some destructors should always be called. Others should only
      * be called if the pipeline is exiting with an error code. */
