+++ src/HexagonOffload.h
@@ -14,9 +14,7 @@ namespace Internal {
 
 /** Pull loops marked with the Hexagon device API to a separate
  * module, and call them through the Hexagon host runtime module. */
-Stmt inject_hexagon_rpc(Stmt s, const Target &host_target, Module &module);
-
-Buffer<uint8_t> compile_module_to_hexagon_shared_object(const Module &device_code);
+Stmt inject_hexagon_rpc(Stmt s, const Target &host_target);
 
 }
 }
