+++ src/Module.cpp
@@ -7,7 +7,6 @@
 #include "CodeGen_C.h"
 #include "CodeGen_Internal.h"
 #include "Debug.h"
-#include "HexagonOffload.h"
 #include "LLVM_Headers.h"
 #include "LLVM_Output.h"
 #include "LLVM_Runtime_Linker.h"
@@ -112,7 +111,6 @@ struct ModuleContents {
     Target target;
     std::vector<Buffer<>> buffers;
     std::vector<Internal::LoweredFunc> functions;
-    std::vector<Module> submodules;
 };
 
 template<>
@@ -169,12 +167,9 @@ const std::vector<Internal::LoweredFunc> &Module::functions() const {
     return contents->functions;
 }
 
+
 std::vector<Internal::LoweredFunc> &Module::functions() {
     return contents->functions;
-}
-
-const std::vector<Module> &Module::submodules() const {
-    return contents->submodules;
 }
 
 Internal::LoweredFunc Module::get_function_by_name(const std::string &name) const {
@@ -193,10 +188,6 @@ void Module::append(const Buffer<> &buffer) {
 
 void Module::append(const Internal::LoweredFunc &function) {
     contents->functions.push_back(function);
-}
-
-void Module::append(const Module &module) {
-    contents->submodules.push_back(module);
 }
 
 Module link_modules(const std::string &name, const std::vector<Module> &modules) {
@@ -225,56 +216,7 @@ Module link_modules(const std::string &name, const std::vector<Module> &modules)
     return output;
 }
 
-Buffer<uint8_t> Module::compile_to_buffer() const {
-    // TODO: This Hexagon specific code should be removed as soon as possible.
-    bool use_shared_object = target().has_feature(Target::HVX_shared_object);
-
-    if (use_shared_object) {
-        return compile_module_to_hexagon_shared_object(*this);
-    } else {
-        llvm::LLVMContext context;
-        std::unique_ptr<llvm::Module> llvm_module(compile_module_to_llvm_module(*this, context));
-
-        llvm::SmallVector<char, 4096> object;
-        llvm::raw_svector_ostream object_stream(object);
-        compile_llvm_module_to_object(*llvm_module, object_stream);
-
-        if (debug::debug_level() >= 2) {
-            debug(2) << "Submodule assembly for " << name() << ": " << "\n";
-            llvm::SmallString<4096> assembly;
-            llvm::raw_svector_ostream assembly_stream(assembly);
-            compile_llvm_module_to_assembly(*llvm_module, assembly_stream);
-            debug(2) << assembly.c_str() << "\n";
-        }
-
-
-        Buffer<uint8_t> result(object.size(), name());
-        memcpy(result.data(), reinterpret_cast<uint8_t*>(&object[0]), object.size());
-        return result;
-    }
-}
-
 void Module::compile(const Outputs &output_files) const {
-    // If there are submodules, recursively lower submodules to
-    // buffers on a copy of the module being compiled, then compile
-    // the copied module.
-    if (!submodules().empty()) {
-        Module lowered_module(name(), target());
-
-        for (const auto &f : functions()) {
-            lowered_module.append(f);
-        }
-        for (const auto &buf : buffers()) {
-            lowered_module.append(buf);
-        }
-        for (const auto &m : submodules()) {
-            auto buf = m.compile_to_buffer();
-            lowered_module.append(buf);
-        }
-        lowered_module.compile(output_files);
-        return;
-    }
-
     if (!output_files.object_name.empty() || !output_files.assembly_name.empty() ||
         !output_files.bitcode_name.empty() || !output_files.llvm_assembly_name.empty() ||
         !output_files.static_library_name.empty()) {
