+++ src/Pipeline.cpp
@@ -3,13 +3,13 @@
 #include "Pipeline.h"
 #include "Argument.h"
 #include "Func.h"
-#include "InferArguments.h"
 #include "IRVisitor.h"
 #include "LLVM_Headers.h"
 #include "LLVM_Output.h"
 #include "Lower.h"
 #include "Outputs.h"
 #include "PrintLoopNest.h"
+#include "WrapExternStages.h"
 
 using namespace Halide::Internal;
 
@@ -40,6 +40,27 @@ Outputs static_library_outputs(const string &filename_prefix, const Target &targ
 }
 
 }  // namespace
+
+/** An inferred argument. Inferred args are either Params,
+ * ImageParams, or Buffers. The first two are handled by the param
+ * field, and global images are tracked via the buf field. These
+ * are used directly when jitting, or used for validation when
+ * compiling with an explicit argument list. */
+struct InferredArgument {
+    Argument arg;
+    Parameter param;
+    Buffer<> buffer;
+
+    bool operator<(const InferredArgument &other) const {
+        if (arg.is_buffer() && !other.arg.is_buffer()) {
+            return true;
+        } else if (other.arg.is_buffer() && !arg.is_buffer()) {
+            return false;
+        } else {
+            return arg.name < other.arg.name;
+        }
+    }
+};
 
 struct PipelineContents {
     mutable RefCount ref_count;
@@ -261,9 +282,159 @@ void Pipeline::compile_to_file(const string &filename_prefix,
     m.compile(outputs);
 }
 
+namespace Internal {
+
+class InferArguments : public IRGraphVisitor {
+public:
+    vector<InferredArgument> &args;
+
+    InferArguments(vector<InferredArgument> &a, const vector<Function> &o, Stmt body)
+        : args(a), outputs(o) {
+        args.clear();
+        for (const Function &f : outputs) {
+            visit_function(f);
+        }
+        if (body.defined()) {
+            body.accept(this);
+        }
+    }
+
+private:
+    vector<Function> outputs;
+    set<string> visited_functions;
+
+    using IRGraphVisitor::visit;
+
+    bool already_have(const string &name) {
+        // Ignore dependencies on the output buffers
+        for (const Function &output : outputs) {
+            if (name == output.name() || starts_with(name, output.name() + ".")) {
+                return true;
+            }
+        }
+        for (const InferredArgument &arg : args) {
+            if (arg.arg.name == name) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+    void visit_exprs(const std::vector<Expr>& v) {
+        for (Expr i : v) {
+            visit_expr(i);
+        }
+    }
+
+    void visit_expr(Expr e) {
+        if (!e.defined()) return;
+        e.accept(this);
+    }
+
+    void visit_function(const Function& func) {
+        if (visited_functions.count(func.name())) return;
+        visited_functions.insert(func.name());
+
+        func.accept(this);
+
+        // Function::accept hits all the Expr children of the
+        // Function, but misses the buffers and images that might be
+        // extern arguments.
+        if (func.has_extern_definition()) {
+            for (const ExternFuncArgument &extern_arg : func.extern_arguments()) {
+                if (extern_arg.is_func()) {
+                    visit_function(Function(extern_arg.func));
+                } else if (extern_arg.is_buffer()) {
+                    include_buffer(extern_arg.buffer);
+                } else if (extern_arg.is_image_param()) {
+                    include_parameter(extern_arg.image_param);
+                }
+            }
+        }
+    }
+
+    void include_parameter(Parameter p) {
+        if (!p.defined()) return;
+        if (already_have(p.name())) return;
+
+        Expr def, min, max;
+        if (!p.is_buffer()) {
+            def = p.get_scalar_expr();
+            min = p.get_min_value();
+            max = p.get_max_value();
+        }
+
+        InferredArgument a = {
+            Argument(p.name(),
+                     p.is_buffer() ? Argument::InputBuffer : Argument::InputScalar,
+                     p.type(), p.dimensions(), def, min, max),
+            p,
+            Buffer<>()};
+        args.push_back(a);
+
+        // Visit child expressions
+        if (!p.is_buffer()) {
+            visit_expr(def);
+            visit_expr(min);
+            visit_expr(max);
+        } else {
+            for (int i = 0; i < p.dimensions(); i++) {
+                visit_expr(p.min_constraint(i));
+                visit_expr(p.extent_constraint(i));
+                visit_expr(p.stride_constraint(i));
+            }
+        }
+    }
+
+    void include_buffer(Buffer<> b) {
+        if (!b.defined()) return;
+        if (already_have(b.name())) return;
+
+        InferredArgument a = {
+            Argument(b.name(), Argument::InputBuffer, b.type(), b.dimensions()),
+            Parameter(),
+            b};
+        args.push_back(a);
+    }
+
+    void visit(const Load *op) {
+        IRGraphVisitor::visit(op);
+        include_parameter(op->param);
+        include_buffer(op->image);
+    }
+
+    void visit(const Variable *op) {
+        IRGraphVisitor::visit(op);
+        include_parameter(op->param);
+        include_buffer(op->image);
+    }
+
+    void visit(const Call *op) {
+        IRGraphVisitor::visit(op);
+        if (op->func.defined()) {
+            Function fn(op->func);
+            visit_function(fn);
+        }
+        include_buffer(op->image);
+        include_parameter(op->param);
+    }
+};
+
+} // namespace Internal
+
 vector<Argument> Pipeline::infer_arguments(Stmt body) {
-    contents->inferred_args = ::infer_arguments(body, contents->outputs);
-    
+
+    user_assert(defined()) << "Can't infer arguments on an undefined Pipeline\n";
+
+    // Infer an arguments vector by walking the IR
+    InferArguments infer_args(contents->inferred_args,
+                              contents->outputs,
+                              body);
+
+    // Sort the Arguments with all buffers first (alphabetical by name),
+    // followed by all non-buffers (alphabetical by name).
+    std::sort(contents->inferred_args.begin(), contents->inferred_args.end());
+
     // Add the user context argument.
     contents->inferred_args.push_back(contents->user_context_arg);
 
@@ -286,6 +457,89 @@ vector<Argument> Pipeline::infer_arguments() {
     return infer_arguments(Stmt());
 }
 
+/** Check that all the necessary arguments are in an args vector. Any
+ * images in the source that aren't in the args vector are returned. */
+vector<Buffer<>> Pipeline::validate_arguments(const vector<Argument> &args, Stmt body) {
+    infer_arguments(body);
+
+    vector<Buffer<>> images_to_embed;
+
+    for (const InferredArgument &arg : contents->inferred_args) {
+
+        if (arg.param.same_as(contents->user_context_arg.param)) {
+            // The user context is always in the inferred args, but is
+            // not required to be in the args list.
+            continue;
+        }
+
+        internal_assert(arg.arg.is_input()) << "Expected only input Arguments here";
+
+        bool found = false;
+        for (Argument a : args) {
+            found |= (a.name == arg.arg.name);
+        }
+
+        if (arg.buffer.defined() && !found) {
+            // It's a raw Buffer used that isn't in the args
+            // list. Embed it in the output instead.
+            images_to_embed.push_back(arg.buffer);
+            debug(1) << "Embedding image " << arg.buffer.name() << "\n";
+        } else if (!found) {
+            std::ostringstream err;
+            err << "Generated code refers to ";
+            if (arg.arg.is_buffer()) {
+                err << "image ";
+            }
+            err << "parameter " << arg.arg.name
+                << ", which was not found in the argument list.\n";
+
+            err << "\nArgument list specified: ";
+            for (size_t i = 0; i < args.size(); i++) {
+                err << args[i].name << " ";
+            }
+            err << "\n\nParameters referenced in generated code: ";
+            for (const InferredArgument &ia : contents->inferred_args) {
+                if (ia.arg.name != contents->user_context_arg.arg.name) {
+                    err << ia.arg.name << " ";
+                }
+            }
+            err << "\n\n";
+            user_error << err.str();
+        }
+    }
+
+    return images_to_embed;
+}
+
+vector<Argument> Pipeline::build_public_args(const vector<Argument> &args, const Target &target) const {
+    // Get all the arguments/global images referenced in this function.
+    vector<Argument> public_args = args;
+
+    // If the target specifies user context but it's not in the args
+    // vector, add it at the start (the jit path puts it in there
+    // explicitly).
+    const bool requires_user_context = target.has_feature(Target::UserContext);
+    bool has_user_context = false;
+    for (Argument arg : args) {
+        if (arg.name == contents->user_context_arg.arg.name) {
+            has_user_context = true;
+        }
+    }
+    if (requires_user_context && !has_user_context) {
+        public_args.insert(public_args.begin(), contents->user_context_arg.arg);
+    }
+
+    // Add the output buffer arguments
+    for (Function out : contents->outputs) {
+        for (Parameter buf : out.output_buffers()) {
+            public_args.push_back(Argument(buf.name(),
+                                           Argument::OutputBuffer,
+                                           buf.type(), buf.dimensions()));
+        }
+    }
+    return public_args;
+}
+
 Module Pipeline::compile_to_module(const vector<Argument> &args,
                                    const string &fn_name,
                                    const Target &target,
@@ -298,56 +552,98 @@ Module Pipeline::compile_to_module(const vector<Argument> &args,
     internal_assert(!new_fn_name.empty()) << "new_fn_name cannot be empty\n";
     // TODO: Assert that the function name is legal
 
-    vector<Argument> lowering_args(args);
+    // TODO: This is a bit of a wart. Right now, IR cannot directly
+    // reference Buffers because neither CodeGen_LLVM nor
+    // CodeGen_C can generate the correct buffer unpacking code.
 
-    // If the target specifies user context but it's not in the args
-    // vector, add it at the start (the jit path puts it in there
-    // explicitly).
-    const bool requires_user_context = target.has_feature(Target::UserContext);
-    bool has_user_context = false;
-    for (Argument arg : lowering_args) {
-        if (arg.name == contents->user_context_arg.arg.name) {
-            has_user_context = true;
-        }
-    }
-    if (requires_user_context && !has_user_context) {
-        lowering_args.insert(lowering_args.begin(), contents->user_context_arg.arg);
-    }
+    // To work around this, we generate two functions. The private
+    // function is one where every buffer referenced is an argument,
+    // and the public function is a wrapper that calls the private
+    // function, passing the global buffers to the private
+    // function. This works because the public function does not
+    // attempt to directly access any of the fields of the buffer.
+
+    Stmt private_body;
+
+    std::vector<std::string> namespaces;
+    std::string simple_new_fn_name = extract_namespaces(new_fn_name, namespaces);
+    string private_name = "__" + simple_new_fn_name;
 
     const Module &old_module = contents->module;
-
-    bool same_compile = !old_module.functions().empty() && old_module.target() == target;
-    // Either generated name or one of the LoweredFuncs in the existing module has the same name.
-    same_compile = same_compile && fn_name.empty();
-    bool found_name = false;
-    for (const auto &lf : old_module.functions()) {
-        if (lf.name == fn_name) {
-            found_name = true;
-            break;
+    if (!old_module.functions().empty() &&
+        old_module.target() == target) {
+        // We can avoid relowering and just reuse the private body
+        // from the old module.
+        for (const LoweredFunc &fn : old_module.functions()) {
+            if (fn.name == private_name) {
+                private_body = fn.body;
+                break;
+            }
         }
-    }
-    same_compile = same_compile && found_name;
-    // Number of args + number of outputs is the same as total args in existing LoweredFunc
-    same_compile = same_compile && (lowering_args.size() + outputs().size()) == old_module.functions().front().args.size();
-    // The initial args are the same.
-    same_compile = same_compile && std::equal(lowering_args.begin(), lowering_args.end(), old_module.functions().front().args.begin());
-    // Linkage is the same.
-    same_compile = same_compile && old_module.functions().front().linkage == linkage_type;
-    // The outputs of a Pipeline cannot change, so no need to test them.
-
-    if (same_compile) {
-        // We can avoid relowering and just reuse the existing module.
         debug(2) << "Reusing old module\n";
-    } else {
+    }
+
+    if (!private_body.defined()) {
         vector<IRMutator *> custom_passes;
         for (CustomLoweringPass p : contents->custom_lowering_passes) {
             custom_passes.push_back(p.pass);
         }
 
-        contents->module = lower(contents->outputs, new_fn_name, target, lowering_args, linkage_type, custom_passes);
+        private_body = lower(contents->outputs, fn_name, target, custom_passes);
     }
 
-    return contents->module;
+    // Get all the arguments/global images referenced in this function.
+    vector<Argument> public_args = build_public_args(args, target);
+
+    vector<Buffer<>> global_images = validate_arguments(public_args, private_body);
+
+    // Create a module with all the global images in it.
+    Module module(simple_new_fn_name, target);
+
+    // Add all the global images to the module, and add the global
+    // images used to the private argument list.
+    vector<Argument> private_args = public_args;
+    for (Buffer<> buf : global_images) {
+        module.append(buf);
+        private_args.push_back(Argument(buf.name(),
+                                        Argument::InputBuffer,
+                                        buf.type(), buf.dimensions()));
+    }
+
+    module.append(LoweredFunc(private_name, private_args,
+                              private_body, LoweredFunc::Internal));
+
+    // Generate a call to the private function, adding an arguments
+    // for the global images.
+    vector<Expr> private_params;
+    for (Argument arg : private_args) {
+        if (arg.is_buffer()) {
+            private_params.push_back(Variable::make(type_of<void*>(), arg.name + ".buffer"));
+        } else {
+            private_params.push_back(Variable::make(arg.type, arg.name));
+        }
+    }
+    string private_result_name = unique_name(private_name + "_result");
+    Expr private_result_var = Variable::make(Int(32), private_result_name);
+    Expr call_private = Call::make(Int(32), private_name, private_params, Call::Extern);
+    Stmt public_body = AssertStmt::make(private_result_var == 0, private_result_var);
+    public_body = LetStmt::make(private_result_name, call_private, public_body);
+
+    module.append(LoweredFunc(new_fn_name, public_args, public_body, linkage_type));
+
+    // Append a wrapper for this pipeline that accepts old buffer_ts
+    // and upgrades them. It will use the same name, so it will
+    // require C++ linkage. We don't need it when jitting.
+    if (!target.has_feature(Target::JIT)) {
+        add_legacy_wrapper(module, module.functions().back());
+    }
+
+    // Also append any wrappers for extern stages that expect the old buffer_t
+    wrap_legacy_extern_stages(module);
+
+    contents->module = module;
+
+    return module;
 }
 
 std::string Pipeline::generate_function_name() const {
