+++ test/correctness/interleave.cpp
@@ -26,8 +26,7 @@ int count_interleaves(Func f) {
     t.set_feature(Target::NoBoundsQuery);
     t.set_feature(Target::NoAsserts);
     f.compute_root();
-    std::vector<Module> submodules;
-    Stmt s = Internal::lower_main_stmt({f.function()}, f.name(), t);
+    Stmt s = Internal::lower({f.function()}, f.name(), t);
     CountInterleaves i;
     s.accept(&i);
     return i.result;
